server/WebScrappingMicroservice/requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
gevent==22.10.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
outcome==1.2.0
pandas==1.5.2
pika==1.3.1
py-mon==1.1.0
pycparser==2.21
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
